fixing after submodule refactoring
